+++ frontend/lib/src/components/elements/Spinner/Spinner.tsx
@@ -28,45 +28,10 @@ import {
   StyledSpinnerTimer,
   ThemedStyledSpinner,
 } from "./styled-components"
+import { formatTime } from "./utils"
 
 export interface SpinnerProps {
   element: SpinnerProto
-}
-
-/**
- * Formats a duration in seconds into a human-readable string.
- *
- * @param seconds - The duration in seconds to format
- * @returns A formatted string representation of the duration in parentheses
- *
- * @example
- * formatTime(1.1)    // "(1.1 seconds)"
- * formatTime(65.3)   // "(1 minute, 5.3 seconds)"
- * formatTime(3661.1) // "(1 hour, 1 minute, 1.1 seconds)"
- *
- * TODO: In the future, we might want to replace this with `Intl.DurationFormat` (see
- * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DurationFormat).
- * But that API is not available on Firefox yet.
- */
-export const formatTime = (seconds: number): string => {
-  const hours = Math.floor(seconds / 3600)
-  const mins = Math.floor((seconds % 3600) / 60)
-  const secs = seconds % 60
-
-  if (hours === 0 && mins === 0) {
-    return `(${secs.toFixed(1)} seconds)`
-  }
-
-  if (hours === 0) {
-    const minText = `${mins} minute${mins === 1 ? "" : "s"}`
-    const secText = secs === 0 ? "" : `, ${secs.toFixed(1)} seconds`
-    return `(${minText}${secText})`
-  }
-
-  const hourText = `${hours} hour${hours === 1 ? "" : "s"}`
-  const minText = mins === 0 ? "" : `, ${mins} minute${mins === 1 ? "" : "s"}`
-  const secText = secs === 0 ? "" : `, ${secs.toFixed(1)} seconds`
-  return `(${hourText}${minText}${secText})`
 }
 
 function Spinner({ element }: Readonly<SpinnerProps>): ReactElement {
